+++ src/app/(normal)/bookmarks/NFTBookmarkList.tsx
@@ -24,7 +24,7 @@ import { useCurrentProfileAll } from '@/hooks/useCurrentProfile.js';
 import { useIsLogin } from '@/hooks/useIsLogin.js';
 import { useNavigatorTitle } from '@/hooks/useNavigatorTitle.js';
 import { FireflySocialMediaProvider } from '@/providers/firefly/SocialMedia.js';
-import type { SimpleHash } from '@/providers/types/SimpleHash.js';
+import type { SimpleHash } from '@/providers/simplehash/type.js';
 
 function getNFTItemContent(nft: SimpleHash.NFT) {
     const chainId = resolveSimpleHashChainId(nft.chain);
