+++ src/helpers/resolveCollectionChain.ts
@@ -2,7 +2,7 @@ import { ChainId } from '@masknet/web3-shared-evm';
 import { first } from 'lodash-es';
 
 import { resolveSimpleHashChainId } from '@/helpers/resolveSimpleHashChain.js';
-import type { SimpleHash } from '@/providers/types/SimpleHash.js';
+import type { SimpleHash } from '@/providers/simplehash/type.js';
 
 export function resolveCollectionChain(collection: SimpleHash.Collection): {
     address: string;
