+++ src/providers/simplehash/formatSimpleHashNFT.ts
@@ -11,7 +11,7 @@ import { resolveNFTImageUrl } from '@/helpers/resolveNFTImageUrl.js';
 import { resolveSimpleHashChainId } from '@/helpers/resolveSimpleHashChain.js';
 import { EVMChainResolver } from '@/mask/bindings/index.js';
 import type { NFTAsset } from '@/providers/types/Firefly.js';
-import type { SimpleHash } from '@/providers/types/SimpleHash.js';
+import type { SimpleHash } from '@/providers/simplehash/type.js';
 
 export const SPAM_SCORE = 50;
