+++ src/providers/simplehash/index.ts
@@ -10,7 +10,7 @@ import { resolveSimpleHashChain } from '@/helpers/resolveSimpleHashChain.js';
 import type { BaseHubOptions } from '@/mask/bindings/index.js';
 import { formatSimpleHashNFT } from '@/providers/simplehash/formatSimpleHashNFT.js';
 import type { NFTAsset } from '@/providers/types/Firefly.js';
-import type { SimpleHash } from '@/providers/types/SimpleHash.js';
+import type { SimpleHash } from '@/providers/simplehash/type.js';
 
 class SimpleHashFactory {
     async getWalletsNFTCollections(
