+++ src/providers/simplehash/type.ts
